refactor(cmd): extract save_config helpers into their own file
